Move predicates and unstructpath to path
And rename unstructpath to selectors, since the package only contains
selectors now. The name of types in the selectors package could be
improved now that the package name is more specific.
