+++ pkg/framework/path/predicates/interface.go
