+++ pkg/framework/path/predicates/map.go
