+++ pkg/framework/path/predicates/number.go
