+++ pkg/framework/path/predicates/slice.go
