+++ pkg/framework/path/predicates/string.go
